11. TelePayBot OOP/requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
mysql-connector==2.2.9
requests==2.23.0
urllib3==1.25.8
